check and configure controllers, utilits
