+++ controllers/system_control.py
@@ -65,12 +65,3 @@ def clean_temp_files():
                     shutil.rmtree(file_path)
         except Exception as e:
             print(f"Не удалось удалить {file_path}. Ошибка: {e}")
-
-
-def check_internet(host="8.8.8.8", port=53, timeout=3):
-    try:
-        socket.setdefaulttimeout(timeout)
-        socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect((host, port))
-        return True
-    except socket.error:
-        return False
