--- utils/open_program.py
@@ -1,8 +1,0 @@
-import subprocess
-
-
-def open_program(path_to_exe: str) -> None:
-    try:
-        subprocess.Popen(path_to_exe)
-    except Exception:
-        pass
